実行プログラムのディレクトリ・ソースファイル名の変更
